feat: Add Modal component to food detail page
The code changes import the Modal component from the Material-UI library and add it to the food detail page. This allows for the display of additional information or actions in a modal dialog. The addition of the Modal component enhances the user experience and provides more flexibility in the design of the page.